+++ corbeans-corda-common/src/main/kotlin/com/github/manosbatsis/corbeans/corda/common/Util.kt
@@ -21,8 +21,6 @@ package com.github.manosbatsis.corbeans.corda.common
 
 import com.fasterxml.jackson.databind.DeserializationFeature
 import com.fasterxml.jackson.dataformat.javaprop.JavaPropsMapper
-import com.github.manosbatsis.corda.rpc.poolboy.config.NodeParams
-import com.github.manosbatsis.corda.rpc.poolboy.config.PoolParams
 import java.util.Properties
 
 object Util {
@@ -75,44 +73,3 @@ class NodesPropertiesLoadingConfigData(
         override val wrappingClass: Class<out NodesPropertiesWrapper>,
         override val ignoreError: Boolean
 ) : NodesPropertiesLoadingConfig
-
-open class NodesProperties : NodesPropertiesWrapper {
-    var cordapPackages: List<String> = mutableListOf()
-    var nodes: Map<String, NodeParams> = mutableMapOf()
-    var primaryControllerType: String? = "com.github.manosbatsis.corbeans.spring.boot.corda.web.CordaSingleNodeController"
-    var notarySpec: TestNotaryProperties = TestNotaryProperties()
-    var flowOverrides: List<String> = mutableListOf()
-    var poolParams: PoolParams = PoolParams()
-    override fun toNodesProperties(): NodesProperties = this
-
-    override fun toString(): String {
-        return "NodesProperties(cordapPackages=$cordapPackages, " +
-                "nodes=$nodes, " +
-                "notarySpec=$notarySpec, " +
-                "flowOverrides=${flowOverrides}), " +
-                "primaryControllerType=${primaryControllerType}, " +
-                "poolParams=${poolParams}"
-    }
-}
-
-
-interface NodesPropertiesWrapper {
-    fun toNodesProperties(): NodesProperties
-}
-
-/**
- * Used to wrap a [NodesProperties] for easier parsing via jackson-dataformats-text,
- * as a temporary workaround to https://github.com/FasterXML/jackson-dataformats-text/issues/100
- */
-class CorbeansNodesPropertiesWrapper : NodesPropertiesWrapper {
-    companion object Config : NodesPropertiesLoadingConfig {
-        override val resourcePath = "/application.properties"
-        override val wrappingClass = CorbeansNodesPropertiesWrapper::class.java
-        override val ignoreError = true
-    }
-
-    var corbeans: NodesProperties? = null
-    override fun toNodesProperties() = corbeans!!
-}
-
-