+++ corbeans-spring-boot-bnms-starter/src/main/kotlin/com/github/manosbatsis/corbeans/spring/boot/corda/bnms/web/support/CorbeansBmnsBaseController.kt
@@ -20,12 +20,11 @@
 package com.github.manosbatsis.corbeans.spring.boot.corda.bnms.web.support
 
 import com.github.manosbatsis.corbeans.spring.boot.corda.bnms.service.CordaBnmsService
+import com.github.manosbatsis.corbeans.spring.boot.corda.config.CordaNodesProperties
 import com.github.manosbatsis.corbeans.spring.boot.corda.web.CorbeansBaseController
-import com.github.manosbatsis.corda.rpc.poolboy.config.NodeParams
 import org.slf4j.LoggerFactory
 import org.springframework.beans.factory.annotation.Autowired
 import java.util.Optional
-import javax.annotation.PostConstruct
 
 open class CorbeansBmnsBaseController : CorbeansBaseController() {
 
@@ -33,37 +32,29 @@ open class CorbeansBmnsBaseController : CorbeansBaseController() {
         private val logger = LoggerFactory.getLogger(CorbeansBmnsBaseController::class.java)
     }
 
-    /** BMNS services by configured name */
-    @Autowired(required = false)
-    lateinit var bmnsServices: Map<String, CordaBnmsService<*>>
+    /** Load config from application.properties */
+    @Autowired
+    lateinit var cordaNodesProperties: CordaNodesProperties
+
+
+    val bnmsServiceType: Class<out CordaBnmsService<*>> by lazy {
+        val typeName = cordaNodesProperties.bnmsServiceType
+                ?: error("A corbeans.nodes.bnmsServiceType property is required")
+        try {
+            Class.forName(typeName) as Class<out CordaBnmsService<*>>
+        }
+        catch (e: Exception){throw RuntimeException(e)}
+    }
 
     /** Store the default node name */
     protected lateinit var defaultNodeName: String
 
-    @PostConstruct
-    fun postConstruct() {
-        logger.debug("Executing postConstruct")
-        if (!::bmnsServices.isInitialized) {
-            logger.warn("No CordaBnmsService beans found")
-            bmnsServices = emptyMap()
-        }
-        // if single node config, use the only node name as default, else reserve explicitly for cordform
-        defaultNodeName = if (bmnsServices.keys.size == 1
-                || !bmnsServices.keys.contains(NodeParams.NODENAME_CORDFORM))
-            bmnsServices.keys.first().replace("NodeService", "")
-        else NodeParams.NODENAME_CORDFORM
-        logger.debug("Auto-configured BMNS services for Corda nodes:: {}, default node: {}",
-                bmnsServices.keys, defaultNodeName)
-    }
-
     /**
-     * Get a BMNS service by name. Default is either the only node name if single,
+     * Get a BMNS service by node name. Default is either the only node name if single,
      * or `cordform` based on node.conf otherwise
      */
     protected fun getBmnsService(optionalNodeName: Optional<String>): CordaBnmsService<*> {
-        val nodeName = networkService.resolveNodeName(optionalNodeName)
-        return this.bmnsServices["${nodeName}BmnsService"]
-                ?: throw IllegalArgumentException("Node not found: `$optionalNodeName`")
+        return networkService.getService(bnmsServiceType, optionalNodeName)
     }
 
 }